Fixed bug where HW view would crash randomly - LILIAN PLEASE READ THE DOCUMENTATION RATHER THAN DOING WHATEVER THE HECK YOU WERE DOING BEFORE JESUS CHRIST DO YOU KNOW HOW LONG IT TOOK ME TO FIND THE BUG WHY WOULD YOU INCREMENT THE "ITEM" COUNTER EVERY TIME THE FUNCTION WAS CALLED RATHER THAN USING INDEXPATH.ROW LIKE LITERALLY EVERYONE ELSE ON THE PLANET AAHHHHHHHHHHHHH

+++ TimetableTableViewController.m
@@ -69,7 +69,7 @@
 
 - (void) tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath{
     
-    int index = indexPath.row;
+    NSInteger index = indexPath.row;
     
     TimetableExpandedViewController *tevc = [[TimetableExpandedViewController alloc] initWithNibName:@"TimetableExpandedViewController" bundle:nil];
